requirements.txt
absl-py==0.8.0
astor==0.8.0
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
Click==7.0
cycler==0.10.0
enum34==1.1.6
Flask==1.1.1
Flask-Cors==3.0.8
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.3.0
gast==0.3.2
google-pasta==0.1.7
grpcio==1.23.0
gunicorn==19.9.0
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
nose==1.3.7
numpy==1.16.4
opencv-python==4.1.0.25
opt-einsum==2.3.2
Pillow==6.2.1
protobuf==3.9.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
scipy==1.2.2
six==1.12.0
subprocess32==3.5.4
tb-nightly==1.15.0a20190806
tensorboard==1.14.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019080601
Werkzeug==0.15.5
wrapt==1.11.2
